Add ObjWithOptional
This commit adds support for TypeScript 4.4 new flag
`exactOptionalPropertyTypes`.

With this flag turned on, you can mark some properties
as optional:

```ts
ObjWithOptional({
    a: NUMBER,
}, {
    b: STRING
})
// This schema will typecheck values as:
{
    a: number,
    b?: string
}

// If you want b to possibly be undefined you can do instead:
ObjWithOptional({
    a: NUMBER,
}, {
    b: MaybeUndefined(STRING)
})
// This will typecheck values as:
{
    a: number,
    b?: string | undefined
}
```

Note that typesafe-schema does not take into account whether the flag
is turned on or not. This means that even though the flag can be set
to true typesafe-schema will always behave as it was set to true.

This should be fine though as it means you should always specify your
optional props as `MaybeUndefined(<prop type>)`.